fetch/metadata/generated/cors-preflight-sec-fetch-dest.sub.html

<!DOCTYPE html>
<!-- This test was procedurally generated. Please do not modify it directly. -->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Dest</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <body>
  <script>
  'use strict';

  promise_test(() => {
    const path = '/fetch/metadata/resources/record-headers.py?key={{uuid()}}';
    return fetch(
        'https://{{hosts[alt][]}}:{{ports[https][0]}}' + path + '&requireHEAD',
        { method: 'DELETE' }
      )
      // This request is expected to fail
      .catch(() => {})
      .then(() => fetch(path + '&retrieve'))
      .then((response) => response.text())
      .then((text) => JSON.parse(text))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-dest');
          assert_equals(headers['sec-fetch-dest'], 'empty');
        });
  }, 'sec-fetch-dest - ');
  </script>
  </body>
</html>
